refactor(fix): split language-agnostic fixers into src/base (#303)
#282 moved the language-agnostic checks into src/base/; the fixers
never followed, so on a Swift repo `doctor` reported the base findings
and `fix` returned `unsupported` for every one of them.

Move the 12 fixers whose output has nothing JS-specific in it
(editorconfig, dependabot, renovate, codeql, github-settings,
codeowners, community-health, ai, claude-skill, cursor-rules,
copilot-instructions, agents-md) into BASE_FIXERS, and have
fixersForLanguage() return [...BASE_FIXERS, ...languageFixers].

Two are now language-aware rather than JS-hardcoded, resolving the
module from the repo's marker files:

- dependabot writes the language's `package-ecosystem` (npm/swift/pip;
  null for Perl drops the manifest block, leaving github-actions)
- codeql uses the module's codeqlLanguages, which makes swift-codeql
  redundant — dropped

github-actions/gitlab-ci/lockfile stay per-module: their content is
language-shaped. Swift gains a swift-lockfile fixer recording the
swift-library preset config, having no package.json to infer from.

Swift .gitignore generation moves to languages/swift/gitignore.ts —
scaffold.ts importing it from fixers.ts closed an import cycle
(lockfile -> setup-presets -> scaffold -> fixers -> lockfile) once the
fixers read the lockfile, which crashed the built CLI at startup.

Closes #303
