+++ src/base/fixers.ts
@@ -1,10 +1,29 @@
 /**
- * The fixer contract, shared by every language module (#286).
+ * The fixer contract plus the language-agnostic fixers (#286, #303).
  *
- * Only the types live here — the fixers themselves belong to their language
- * module. `fix` concatenates the set for the repo's detected language.
+ * These are the fixers for the checks in ./checks.ts — repo hygiene, security
+ * automation, community health, AI agent files, GitHub repo settings. Nothing
+ * here reads a package.json or emits a toolchain-specific step, so every
+ * language module gets them for free; `fix` concatenates them with the module's
+ * own set.
+ *
+ * Not here: `github-actions` / `gitlab-ci` / `lockfile`. Their *content* is
+ * language-shaped (CI steps, recorded tool choices), so each module ships its
+ * own — see src/base/ci.ts for the shell they share.
  */
+import { installAgentRules, installAiSetup } from '../cli/generators/agent-rules.js'
+import { generateCommunityHealth } from '../cli/generators/community-health.js'
+import { generateCodeowners, generateEditorConfig } from '../cli/generators/misc.js'
+import {
+	generateCodeQLWorkflow,
+	generateDependabotConfig,
+	generateRenovateConfig,
+} from '../cli/generators/security.js'
+import { copyPreset } from '../cli/utils/copy-preset.js'
+import { detectLanguage } from '../cli/utils/detect-language.js'
 import type { Lockfile } from '../cli/utils/lockfile.js'
+import { resolveLanguageModule } from '../languages/registry.js'
+import { applyGithubSettings } from './github-settings.js'
 import type { CheckResult } from './types.js'
 
 /** A parsed package.json, or null when the repo has none (any non-JS repo). */
@@ -36,3 +55,180 @@ export interface Fixer {
 	canFixDrift?: boolean
 	run(ctx: FixerContext): Promise<{ filesWritten: string[] }>
 }
+
+/** The repo's language module, resolved from its marker files. */
+async function moduleFor(targetDir: string) {
+	return resolveLanguageModule(await detectLanguage(targetDir))
+}
+
+export const BASE_FIXERS: Fixer[] = [
+	{
+		target: 'editorconfig',
+		description: 'Scaffold .editorconfig (UTF-8, LF, tab indent)',
+		appliesTo: ['EditorConfig'],
+		outputs: ['.editorconfig'],
+		canFixDrift: true,
+		async run({ targetDir }) {
+			await generateEditorConfig(targetDir)
+			return { filesWritten: ['.editorconfig'] }
+		},
+	},
+	{
+		target: 'dependabot',
+		description:
+			'Scaffold the canonical .github/dependabot.yml (monthly, grouped: production-minor/dev-minor/major-updates) + the dependabot-automerge workflow',
+		appliesTo: ['Dependabot'],
+		outputs: ['.github/dependabot.yml', '.github/workflows/dependabot-automerge.yml'],
+		canFixDrift: true,
+		async run({ targetDir }) {
+			const { dependabotEcosystem } = await moduleFor(targetDir)
+			return { filesWritten: await generateDependabotConfig(targetDir, dependabotEcosystem) }
+		},
+	},
+	{
+		target: 'renovate',
+		description: 'Scaffold renovate.json (weekly schedule; alternative to Dependabot)',
+		appliesTo: ['Dependabot'],
+		outputs: ['renovate.json'],
+		riskLevel: 'safe-add',
+		async run({ targetDir }) {
+			await generateRenovateConfig(targetDir)
+			return { filesWritten: ['renovate.json'] }
+		},
+	},
+	{
+		target: 'codeql',
+		description: 'Scaffold .github/workflows/codeql.yml (security scanning)',
+		appliesTo: ['CodeQL'],
+		outputs: ['.github/workflows/codeql.yml'],
+		async run({ targetDir }) {
+			const { codeqlLanguages } = await moduleFor(targetDir)
+			return { filesWritten: await generateCodeQLWorkflow(targetDir, codeqlLanguages) }
+		},
+	},
+	{
+		target: 'github-settings',
+		description:
+			'Apply branch protection + auto-merge + workflow permissions + code-scanning ruleset on GitHub via gh api (mutates the remote repo, not files)',
+		appliesTo: [
+			'Branch protection',
+			'Merge settings',
+			'Workflow permissions',
+			'Code-scanning gate',
+		],
+		outputs: ['GitHub repo settings (remote, via gh api)'],
+		// safe-add is load-bearing: it exempts this fixer from the `--diff` shadow-run
+		// (previewFixer copies to tmp and *executes* run(), which would fire real
+		// `gh api` PUTs during a mere preview).
+		riskLevel: 'safe-add',
+		canFixDrift: true,
+		async run({ targetDir }) {
+			return { filesWritten: await applyGithubSettings(targetDir) }
+		},
+	},
+	{
+		target: 'codeowners',
+		description: 'Scaffold .github/CODEOWNERS with commented examples',
+		appliesTo: ['CODEOWNERS'],
+		outputs: ['.github/CODEOWNERS'],
+		riskLevel: 'safe-add',
+		canFixDrift: false,
+		async run({ targetDir }) {
+			const written = await generateCodeowners(targetDir)
+			return { filesWritten: [written] }
+		},
+	},
+	{
+		target: 'community-health',
+		description: 'Scaffold CONTRIBUTING.md, SECURITY.md, PR + issue templates',
+		appliesTo: ['Community health'],
+		outputs: [
+			'CONTRIBUTING.md',
+			'SECURITY.md',
+			'.github/PULL_REQUEST_TEMPLATE.md',
+			'.github/ISSUE_TEMPLATE/bug_report.md',
+			'.github/ISSUE_TEMPLATE/feature_request.md',
+		],
+		riskLevel: 'safe-add',
+		canFixDrift: false,
+		async run({ targetDir }) {
+			const filesWritten = await generateCommunityHealth(targetDir)
+			return { filesWritten }
+		},
+	},
+	{
+		target: 'ai',
+		description:
+			'Install all AI agent files at once (AGENTS.md, CLAUDE.md, Cursor, Copilot, Claude skill, MCP example)',
+		appliesTo: ['AI setup'],
+		outputs: [
+			'AGENTS.md',
+			'CLAUDE.md',
+			'.cursor/rules/repo-tooling.mdc',
+			'.github/copilot-instructions.md',
+			'.claude/skills/repo-tooling.md',
+			'.mcp.json.example',
+			// Only written when the repo ships its own skills/<name>/SKILL.md.
+			'README.md',
+		],
+		// Every output is a delimited-block upsert or a `.example` file — existing
+		// user content is never clobbered.
+		riskLevel: 'safe-merge',
+		canFixDrift: true,
+		async run({ targetDir }) {
+			const filesWritten = await installAiSetup(targetDir)
+			return { filesWritten }
+		},
+	},
+	{
+		target: 'claude-skill',
+		description: 'Install the repo-tooling Claude Code skill into .claude/skills/',
+		appliesTo: ['Claude skill'],
+		outputs: ['.claude/skills/repo-tooling.md'],
+		riskLevel: 'safe-add',
+		canFixDrift: true,
+		async run({ targetDir }) {
+			const result = await copyPreset('claude-skill', targetDir)
+			return { filesWritten: [result.target] }
+		},
+	},
+	{
+		target: 'cursor-rules',
+		description: 'Install the repo-tooling rules for Cursor (.cursor/rules/repo-tooling.mdc)',
+		appliesTo: ['Cursor rules'],
+		outputs: ['.cursor/rules/repo-tooling.mdc'],
+		riskLevel: 'safe-add',
+		canFixDrift: true,
+		async run({ targetDir }) {
+			const written = await installAgentRules(targetDir, 'cursor')
+			return { filesWritten: [written] }
+		},
+	},
+	{
+		target: 'copilot-instructions',
+		description:
+			'Install the repo-tooling rules for GitHub Copilot (.github/copilot-instructions.md)',
+		appliesTo: ['Copilot instructions'],
+		outputs: ['.github/copilot-instructions.md'],
+		// Upserts a delimited block — never clobbers the consumer's own instructions.
+		riskLevel: 'safe-merge',
+		canFixDrift: true,
+		async run({ targetDir }) {
+			const written = await installAgentRules(targetDir, 'copilot')
+			return { filesWritten: [written] }
+		},
+	},
+	{
+		target: 'agents-md',
+		description: 'Install the repo-tooling rules into AGENTS.md (universal agent instructions)',
+		appliesTo: ['AGENTS.md rules'],
+		outputs: ['AGENTS.md'],
+		// Upserts a delimited block — never clobbers existing AGENTS.md content.
+		riskLevel: 'safe-merge',
+		canFixDrift: true,
+		async run({ targetDir }) {
+			const written = await installAgentRules(targetDir, 'agents-md')
+			return { filesWritten: [written] }
+		},
+	},
+]
